refactor: move utxo coins to utxo.ts
Ticket: BTC-1092
