+++ CODEOWNERS
@@ -81,7 +81,7 @@
 /modules/sdk-unified-wallet @BitGo/ethalt-team
 /modules/sjcl/ @BitGo/coins @BitGo/custody-experience @BitGo/wallet-platform
 /modules/statics/ @BitGo/custody-experience @BitGo/wallet-platform @BitGo/ethalt-team
-/modules/statics/src/utxo @BitGo/btc-team
+/modules/statics/src/utxo.ts @BitGo/btc-team
 /modules/web-demo/ @BitGo/coins @BitGo/custody-experience @BitGo/wallet-platform
 /scripts/ @BitGo/coins @BitGo/custody-experience @BitGo/wallet-platform
 /types/ @BitGo/coins @BitGo/custody-experience @BitGo/wallet-platform
